send message test
