+++ src/main/scala/ohnosequences/saws/signing/sqsbench/SQSPerf.scala
@@ -1,0 +1,5 @@
+package ohnosequences.saws.signing.sqsbench
+
+object SQSPerf {
+
+}
